refactor: 1. 불필요한 GameResult 에 대한 List 감싸기를 제거 2. GameResult 내 불필요한 깊은복사 제거 3. GameFinalResult 를 GameResponse로 변경하여 UI용 객체임을 명시적으로 표시
