--- src/main/java/racingGame/response/GameFinalResult.java
@@ -1,8 +1,0 @@
-package racingGame.response;
-
-import java.util.List;
-import racingGame.model.Car;
-
-public record GameFinalResult(List<GameResult> gameResult, List<Car> winners) {
-    
-}
