+++ src/test/java/racingGame/response/GameResultTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingGame.model.Car;
+import racingGame.model.GameResult;
 import racingGame.model.ProgressRecord;
 
 class GameResultTest {
